data/producer/requirements.txt
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
kafka-python==2.3.0
numpy==2.3.5
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
schedule==1.2.2
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
